setup.cppにclient変数を定義、SSL認証を無効化を行い、他ではexternする感じにしました
