+++ src/services/transcription/whisper_client.h
